FeRAM Starting Template version 1.00
Cleaned up the template to create a true starting point for the FeRAM
code. It only contains the feram_i2c header file that prepares the data
based on the memory IC's characteristics (such as the page write cycle).
I still need to make the template for SPI. I also updated the "readme"
file.