feram_i2c.h update
Verifying to see if the feram_write and the feram_read functions work.
Created lines of code inserting data into the MR44V064A and reading from
it. I also tried to see if the data remained within the IC when turned
off in order to verify it worked. I had a hard time working with the
DTU8 debugger to check to see if the functions worked. In order to
verify if the code was working, I began to use the "Watch" function to
monitor the SFR used to read and write on the FeRAM IC. Based on the
results entered into the SFR, it seems that the code is working (I2C
firmware)